+++ module/documents/combat.mjs
@@ -4,13 +4,12 @@
  */
 
 export class CraskCombat extends Combat {
-    /** @override */
-    prepareData() {
-      // Prepare data for the actor. Calling the super version of this executes
-      // the following, in order: data reset (to clear active effects),
-      // prepareBaseData(), prepareEmbeddedDocuments() (including active effects),
-      // prepareDerivedData().
-      super.prepareData();
-    }
+    
+  _sortCombatants(a, b) {
+    console.log("SORTING!");
+    let phase = a.combat.phase;
+    return super._sortCombatants(b, a);
+  }
+
 
 }